refactor: move State to base refactor: move StateMachine to base refactor: minor consistency names add: define render data structures
