+++ requirements.txt
@@ -1,3 +1,4 @@
+colorama==0.4.6
 langchain==0.0.348
 langchain-core==0.0.12
 langsmith==0.0.69
@@ -7,3 +8,4 @@ pydantic_core==2.14.5
 pygame==2.5.2
 python-dotenv==1.0.0
 pytest
+termcolor==2.4.0
